Rename "premade" (the meaning of which was ambiguous) to "premade_models" (clearer).
PiperOrigin-RevId: 425391668
