+++ keras/api/BUILD
@@ -105,8 +105,8 @@ keras_packages = [
     "keras.optimizer_v2.optimizer_v2",
     "keras.optimizer_v2.rmsprop",
     "keras.optimizers",
-    "keras.premade.linear",
-    "keras.premade.wide_deep",
+    "keras.premade_models.linear",
+    "keras.premade_models.wide_deep",
     "keras.preprocessing.image",
     "keras.preprocessing.sequence",
     "keras.preprocessing.text",
