+++ keras/distribute/keras_premade_models_test.py
@@ -20,8 +20,8 @@ from keras.engine import sequential
 from keras.layers import core
 from keras.optimizer_v2 import adagrad
 from keras.optimizer_v2 import gradient_descent
-from keras.premade import linear
-from keras.premade import wide_deep
+from keras.premade_models import linear
+from keras.premade_models import wide_deep
 from keras.utils import dataset_creator
 import numpy as np
 import tensorflow.compat.v2 as tf
