+++ keras/premade_models/linear.py
