+++ keras/premade_models/linear_test.py
@@ -26,7 +26,7 @@ from keras.engine import training
 from keras.feature_column import dense_features_v2
 from keras.layers import core
 from keras.optimizer_v2 import gradient_descent
-from keras.premade import linear
+from keras.premade_models import linear
 
 
 @keras_parameterized.run_all_keras_modes(always_skip_v1=True)
